+++ cognite/client/data_classes/assets.py
@@ -19,6 +19,7 @@ class Asset(CogniteResource):
         created_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
         last_updated_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
         root_id (int): A JavaScript-friendly internal ID for the object.
+        aggregates (Dict[str, Any]): Aggregated metrics of the asset
         parent_external_id (str): The external ID provided by the client. Must be unique within the project.
         cognite_client (CogniteClient): The client to associate with this object.
     """
@@ -35,6 +36,7 @@ class Asset(CogniteResource):
         created_time: int = None,
         last_updated_time: int = None,
         root_id: int = None,
+        aggregates: Dict[str, Any] = None,
         parent_external_id: str = None,
         cognite_client=None,
     ):
@@ -48,6 +50,7 @@ class Asset(CogniteResource):
         self.created_time = created_time
         self.last_updated_time = last_updated_time
         self.root_id = root_id
+        self.aggregates = aggregates
         self.parent_external_id = parent_external_id
         self._cognite_client = cognite_client
 
@@ -255,7 +258,7 @@ class AssetFilter(CogniteFilter):
         source (str): The source of the asset.
         created_time (Dict[str, Any]): Range between two timestamps.
         last_updated_time (Dict[str, Any]): Range between two timestamps.
-        root (bool): Whether the filtered assets are root assets, or not.
+        root (bool): Whether the listed assets are root assets, or not.
         external_id_prefix (str): The external ID provided by the client. Must be unique within the project.
         cognite_client (CogniteClient): The client to associate with this object.
     """