+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Changes are grouped as follows
 
 ## [Unreleased]
 
+### Changed
+- `Asset` class now has an `aggregates` property.
+
 ## [1.1.7] - 2019-09-13
 ### Fixed
 - `testing.mock_cognite_client()` so that it still accepts arguments after exiting from mock context.